added missing cast

+++ filter/merge_replay_matrix.h
@@ -74,7 +74,7 @@ typedef struct {
 extern "C" {
 #endif
 
-#define compute_WN(mat) ((mat)->rem_nrows * (mat)->weight)
+#define compute_WN(mat) ((double) (mat)->rem_nrows * (double) (mat)->weight)
 #define compute_WoverN(mat) (((double)(mat)->weight)/((double)(mat)->rem_nrows))
 
 void initMat(filter_matrix_t *, int, uint32_t, uint32_t);
